backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]